// src/generated/api/admin/models/adminUserListItemResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aller Meal API
 * Aller Meal Release 1 public, member, and admin API contract.
 * OpenAPI spec version: 1.0.0
 */
import type { AdminUserListItemResponseEmailVerificationStatus } from "./adminUserListItemResponseEmailVerificationStatus";
import type { AdminUserListItemResponseRole } from "./adminUserListItemResponseRole";
import type { AdminUserListItemResponseStatus } from "./adminUserListItemResponseStatus";

export interface AdminUserListItemResponse {
  userId: string;
  email: string;
  role: AdminUserListItemResponseRole;
  status: AdminUserListItemResponseStatus;
  emailVerificationStatus: AdminUserListItemResponseEmailVerificationStatus;
  createdAt: string;
}