// src/generated/api/admin/models/listAdminUsersStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aller Meal API
 * Aller Meal Release 1 public, member, and admin API contract.
 * OpenAPI spec version: 1.0.0
 */

export type ListAdminUsersStatus =
  (typeof ListAdminUsersStatus)[keyof typeof ListAdminUsersStatus];

export const ListAdminUsersStatus = {
  ACTIVE: "ACTIVE",
  WITHDRAWAL_PENDING: "WITHDRAWAL_PENDING",
  SUSPENDED: "SUSPENDED",
} as const;